Huge update - very customizable after feedback from users
v1.0.20160430 - Main and Custom switches can now be momentary or have
on/off states. GPIO pin numbers for Main/Custom is now configurable.
Ability to force Echo/Alexa to be restricted in controlling the Main
switch when running ON/OFF commands when configured to be stateful
ON/OFF (not momentary). Refresh works and updates Main/Custom tile per
the GPIO pin status. Version contains lots of enhancements, bug fixes
and future proofing.